// types/enum.ts
export enum Grade {
    P = "grade_p",
    SS = "grade_ss",
    S = "grade_s",
    A = "grade_a",
    B = "grade_b",
    C = "grade_c",
    D = "grade_d",
    F = "grade_f",
  }

export enum Tier {
    BronzeI = "Bronze I",
    BronzeII = "Bronze II",
    BronzeIII = "Bronze III",
    BronzeIV = "Bronze IV",
    Challenger = "Challenger",
    DiamondI = "Diamond I",
    DiamondII = "Diamond II",
    DiamondIII = "Diamond III",
    DiamondIV = "Diamond IV",
    GoldI = "Gold I",
    GoldII = "Gold II",
    GoldIII = "Gold III",
    GoldIV = "Gold IV",
    Grandmaster = "Grandmaster",
    Master = "Master",
    PlatinumI = "Platinum I",
    PlatinumII = "Platinum II",
    PlatinumIII = "Platinum III",
    PlatinumIV = "Platinum IV",
    SilverI = "Silver I",
    SilverII = "Silver II",
    SilverIII = "Silver III",
    SilverIV = "Silver IV",
    Unranked = "Unranked",
}

export type Progress = "P" | "SS" | "S" | "A" | "B" | "C" | "D" | "F";